+++ src/index.js
@@ -1,3 +1,3 @@
 export * from './convert'
 export * from './geo'
-export * from './cpa'
+export { cpa } from './cpa'