+++ test/index.test.js
@@ -1,7 +1,7 @@
 import test from 'ava'
 import * as geo from '../src/index'
 
-test ('contains all functions of the library', t => {
+test ('index contains all functions of the library', t => {
   const allFunctions = [ 
     'degToRad',
     'radToDeg',
@@ -9,6 +9,7 @@ test ('contains all functions of the library', t => {
     'meterPerSecondToKnots',
     'knotsToKmPerHour',
     'kmPerHourToKnots',
+
     'EARTH_RADIUS',
     'isLatLon',
     'isLatLng',
@@ -34,7 +35,9 @@ test ('contains all functions of the library', t => {
     'normalizeLongitude',
     'normalizeLocation',
     'average',
-    'getBoundingBox' 
+    'getBoundingBox',
+
+    'cpa',
   ]
 
   t.deepEqual(Object.keys(geo).sort(), allFunctions.sort())