+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 # not yet released, version 0.4.0
 
 - Implemented `getBoundingBox`, `insideBoundingBox`, `insideCircle`, `insidePolygon`, `average`.
+- Implemented `cpa` (closest point of approach)
 - Renamed `angle` to `heading` everywhere in the library (properties and functions). Breaking change.